Removed question views from auth_views and placed them in the correctly named question_views folder
